tests/requirements-test.txt
# Testing dependencies for UDA Testing Framework
requests>=2.31.0
python-socketio>=5.8.0
asyncio-mqtt>=0.16.1
aiohttp>=3.9.0
pydantic>=2.5.0